@@ -27,12 +27,12 @@ pyright = ">=1.1.350,<1.2"
 
 [tool.ruff]
 line-length = 115
-target-version = "py38"
+target-version = "py39"
 output-format = "grouped"
 
 [tool.ruff.lint]
 select = ["F","E","W","C","I","N","UP","ANN","S","B","A","COM","C4","T20","PT","ARG","TD","RUF"]
-ignore = ["A001","A002","A003","ANN101","ANN102","ANN401","C901","N8","B008","F405","F821"]
+ignore = ["A001","A002","A003","ANN401","C901","N8","B008","F405","F821"]
 
 [tool.ruff.lint.per-file-ignores]
 "__init__.py" = ["F401"]
